remove MetadataLocation and associated magic
MetadataLocation was using FileWatchers to automatically refresh IdP
information. This should happen rarely if at all, and causes issues when
metadata fails to parse. Also removed automatic gathering of metadata,
which may come back at some point (OpenId connect does something like
that),
but it should not be writing to the filesystem as the system likely
doesn't
have permissions to do that.