fully working SP-Initiated Redirect binding
According to this older post, the cookie
middleware needs to be set to active mode. On SO the default is supposed
to be active, but it doesn't appear to be.

Combined with a redirectUri temporary property, we can get SP-initiated
sign-in through testshib.org and view all the claims.

Note that Firefox is best for testing. HttpListener (I think) is doing
something weird with the transfers over the wire that might be Https
specific. Chrome doesn't seem to like it, Firefox is ok but looks wierd,
and IE actually works but I didn't even think to check until writing
this!

http://leastprivilege.com/2014/02/21/test-driving-the-ws-federation-authentication-middleware-for-katana/
http://stackoverflow.com/a/26277965/113225